Make OpenMP Dynamic for Blender Plugin Linking
Prior to this commit OpenMP pointed to a static library causing the
Blender plugin to crash the program due to OpenMP throwing a trap for
initializing multiple OpenMP instances. This commit partially fixes
that by changing the library to be dynamicly linked so that the same
blender OpenMP library can be used by the plugin but the python API
may also use an OpenMP library without Blender.