Replace TravisCI with Azure Pipelines
Prior to this commit Linux and Windows platforms had been converted
from TravisCI to Azure Pipelines allowing for tests to be run in the
CI system but MacOS had not been converted yet. This commit fixes that
by removing the TravisCI files and updating the Azure Pipeline process
to include MacOS build and tests.